Renamed mock package and go file inside openlibrary->mock
